+++ src/migrations/20240318110545_add_isdeleted_to_bookmarks.ts
@@ -7,7 +7,11 @@ export async function up(knex: Knex): Promise<void> {
 }
 
 export async function down(knex: Knex): Promise<void> {
-  return knex.schema.table('bookmarks', function (table) {
-    table.dropColumn('isdeleted');
-  });
+  const exits = await knex.schema.hasTable('bookmarks');
+
+  if (exits) {
+    return knex.schema.table('bookmarks', function (table) {
+      table.dropColumn('isdeleted');
+    });
+  }
 }
